move core logic of client fixtures into util
