+++ localstack/testing/pytest/fixtures.py
@@ -24,7 +24,6 @@ from werkzeug import Request, Response
 
 from localstack import config, constants
 from localstack.aws.accounts import get_aws_account_id
-from localstack.aws.connect import ExternalAwsClientFactory, ExternalClientFactory
 from localstack.constants import TEST_AWS_ACCESS_KEY_ID, TEST_AWS_SECRET_ACCESS_KEY
 from localstack.services.stores import (
     AccountRegionBundle,
@@ -52,30 +51,6 @@ LOG = logging.getLogger(__name__)
 
 # URL of public HTTP echo server, used primarily for AWS parity/snapshot testing
 PUBLIC_HTTP_ECHO_SERVER_URL = "http://httpbin.org"
-
-
-@pytest.fixture(scope="session")
-def aws_session():
-    return boto3.Session()
-
-
-@pytest.fixture(scope="session")
-def aws_client_factory(aws_session):
-    config = None
-    if os.environ.get("TEST_DISABLE_RETRIES_AND_TIMEOUTS"):
-        config = botocore.config.Config(
-            connect_timeout=1_000, read_timeout=1_000, retries={"total_max_attempts": 1}
-        )
-
-    if os.environ.get("TEST_TARGET") == "AWS_CLOUD":
-        return ExternalAwsClientFactory(session=aws_session, config=config)
-    else:
-        return ExternalClientFactory(session=aws_session, config=config)
-
-
-@pytest.fixture(scope="session")
-def aws_client(aws_client_factory):
-    return aws_client_factory()
 
 
 def _resource(service):
